Add a CLI to the gem
This commit adds a functioning CLI with a minimal output formatter.
Adding additional formatters should be straightforward, and only
requires inheriting from AbstractFormatter and perhaps defining the
formatter as a possibility under Options (though this latter part is not
necessary).

The CLI also comes with associated tests, as well as tests for many
other areas of the code. As part of this, I moved around the directory
structure of the spec folder -- I couldn't think of any reasons to
maintain /lib/ in the path.

I have also changed the base debsecan.rb to stop using autoload. It is
generally frowned on in the ruby community to use autoload syntax, so
the file is now just the set of requires for the whole project. Some of
these, specifically some of the gem requires, may not be strictly necessary,
and can likely be removed in a later commit. I did, however, take the
opportunity to remove the use of Psych in favor of the Oj gem.

There are still several spec files to be written or needing more specs:
  - abstract_formatter_spec.rb
  - simple_formatter_spec.rb
  - options_spec.rb

I likely need to write some additional specs for Runner and CLI as well,
as they are fairly sparse right now.

Coverage is at 92.3% as of this commit. I believe this is slightly low,
as simplecov seems to account for files that are not explicitly tested
strangely.