Scanner, handle virtual packages more sanely
When the Scanner is given a Package which describes a source, but the
package set the Scanner is configured with does not contain a Package
for that source, it crashes.

When searching for a Package during a scan, allow either of the other
packet's name or the source to match the current Package. Then, when
collecting defects for comparison, collect all defects for both the
current and source package.

I refactored several methods as part of fixing this, and cleaned up some
unnecessary mocking in the specs.