Fix travis build configuration
Travis runs tests on Ubuntu, while the runner and cli specs assumed that
they were being run on Debian systems. This caused test failures when
the system release was used as a comparison to the fixtures in the
specs. A later commit will provide testing under Ubuntu assumptions as
well.

Additionally, the default Travis image does not install libapt-pkg-dev,
causing installation of the apt-pkg gem to fail.